% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{calc_meanClimDaily_dayOfYearWindow}
\alias{calc_meanClimDaily_dayOfYearWindow}
\title{Calculates the seasonal pattern (i.e. climatological mean)}
\usage{
calc_meanClimDaily_dayOfYearWindow(obs, doy, inc)
}
\arguments{
\item{obs}{is a vector, representing a time series}

\item{doy}{is the day of year for each value in the time series}

\item{inc}{is the half-window size used in moving average}
}
\description{
Calculates the seasonal pattern (i.e. climatological mean)
}
