% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{data_A5030502}
\alias{data_A5030502}
\title{Catchment data for Scott Creek in South Australia for period 1976-1985.}
\format{
A list with 4 elements
\describe{
  \item{times}{Vector of times in POSIXct format}
  \item{P}{Vector of precipitation data (mm)}
  \item{PET}{Vector of PET data (mm) (seasonally variable, no changes annually)}
  \item{Qobs}{Vector of observed streamflow (mm)}
}
}
\usage{
data_A5030502
}
\description{
Catchment data for Scott Creek in South Australia for period 1976-1985.
}
\keyword{datasets}
