% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egClimData}
\alias{egClimData}
\title{Climate attributes from projections.}
\format{
A data frame with 15 rows and 12 variables:
\describe{
  \item{P_day_all_tot_m}{change in mean annual total P, fraction}
  \item{P_day_all_seasRatioMarAug}{change in seasonal (Mar-Aug) ratio of P, fraction}
  \item{P_day_all_P99}{change in 99th percentile of P, fraction}
  \item{Temp_day_all_avg}{change in averageTemp, additive}
  \item{Name}{name of the climate model}
  \item{Avg. Deficit}{performance metric values}
}
}
\usage{
egClimData
}
\description{
A example dataset containing the climate attribute values
in fraction/additive change
}
\keyword{datasets}
