% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egMultiSiteSim}
\alias{egMultiSiteSim}
\title{Output from call to generateScenarios() using multi-site model (see example 5 in generateScenarios).}
\format{
A list with 4 elements
\describe{
  \item{Rep1}{List containing majority of simulation output, including output for different calibration stages }
  \item{simDates}{the dates of the simulation}
  \item{expSpace}{the exposure space of the simulation}
  \item{controlFile}{the setting in the control file}
}
}
\usage{
egMultiSiteSim
}
\description{
Output from call to generateScenarios() using multi-site model (see example 5 in generateScenarios).
}
\keyword{datasets}
