% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egScalSummary}
\alias{egScalSummary}
\title{Summary of a simple scaled scenario.}
\format{
A list containing 3 elements
\describe{
  \item{simDates}{the dates of the simulation}
  \item{expSpace}{the exposure space of the simulation}
  \item{controlFile}{"scaling"}
}
}
\usage{
egScalSummary
}
\description{
Summary generated using the function \code{getSimSummary}.
}
\keyword{datasets}
