\name{sim.stoch}
\alias{sim.stoch}
\docType{data}
\title{
Example perturbed stochastic climates for Scott Creek. 
}
\description{
Based on example in vignette for stochastic simulation.
}
\format{A large list with perturbed climate simulations and metadata}
\source{Generated by generateScenarios() - see example in vignette}
\keyword{datasets}
