% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egSimOATPerformance}
\alias{egSimOATPerformance}
\title{Performance metrics of the tank model using OAT scenarios.}
\format{
A list with 2 elements
\describe{
  \item{Avg. Deficit}{average daily deficit of water, litres}
  \item{Reliability}{reliability of the tank, fraction}
}
}
\usage{
egSimOATPerformance
}
\description{
Performance metrics of the tank model using OAT scenarios.
}
\keyword{datasets}
