% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egSimOATSummary}
\alias{egSimOATSummary}
\title{Summary of a OAT scenario.}
\format{
A list containing 13 elements
}
\usage{
egSimOATSummary
}
\description{
Summary generated using the function \code{getSimSummary} for
a scenarios generated using stochastic models for an OAT exposure space
}
\keyword{datasets}
