% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egSimPerformance_systemB}
\alias{egSimPerformance_systemB}
\title{Performance metrics of an alternate tank model using regGrid scenarios.}
\format{
A list with 2 elements
\describe{
  \item{Avg. Deficit}{average daily deficit of water, litres}
  \item{Reliability}{reliability of the tank, fraction}
}
}
\usage{
egSimPerformance_systemB
}
\description{
Performance metrics of an alternate tank model using regGrid scenarios.
}
\keyword{datasets}
