% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egSimSummary}
\alias{egSimSummary}
\title{Summary of a regGrid scenario.}
\format{
A list containing 13 elements
}
\usage{
egSimSummary
}
\description{
Summary generated using the function \code{getSimSummary} for
a scenarios generated using stochastic models for a regGrid exposure space
}
\keyword{datasets}
