% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{data
exampleData}
\format{
R table or .csv
}
\source{
An example GCAM run
}
\usage{
exampleData
}
\description{
data
exampleData
}
\examples{
\dontrun{
 library(argus);
 argus::exampleData
}
}
\keyword{datasets}
