% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{package}
\name{foreSIGHT}
\alias{foreSIGHT-package}
\alias{foreSIGHT}
\title{foreSIGHT: A package for Systems Insights from Generation of Hydroclimatic Timeseries}
\description{
A tool to create hydroclimate scenarios, stress test systems and visualize
system performance in scenario-neutral climate change impact assessments.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ClimateAnalytics/foreSIGHT/issues}
}

}
\author{
\strong{Maintainer}: David McInerney \email{david.mcinerney@adelaide.edu.au} (\href{https://orcid.org/0000-0003-4876-8281}{ORCID})

Authors:
\itemize{
  \item Bree Bennett \email{bree.bennett@adelaide.edu.au} (\href{https://orcid.org/0000-0002-2131-088X}{ORCID})
  \item Sam Culley \email{sam.culley@adelaide.edu.au} (\href{https://orcid.org/0000-0003-4798-8522}{ORCID})
  \item Anjana Devanand \email{anjana.devanand@adelaide.edu.au} (\href{https://orcid.org/0000-0001-9422-3894}{ORCID})
  \item Seth Westra \email{seth.westra@adelaide.edu.au} (\href{https://orcid.org/0000-0003-4023-6061}{ORCID})
}

Other contributors:
\itemize{
  \item Danlu Guo \email{danlu.guo@adelaide.edu.au} (\href{https://orcid.org/0000-0003-1083-1214}{ORCID}) [contributor]
  \item Holger Maier \email{holger.maier@adelaide.edu.au} (\href{https://orcid.org/0000-0002-0277-6887}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
