% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_avg}
\alias{func_avg}
\title{Calculates average of time series}
\usage{
func_avg(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates average of time series
}
