% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_avgDSD}
\alias{func_avgDSD}
\title{Calculates average dry spell duration (below threshold)}
\usage{
func_avgDSD(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$threshold denoting the threshold}
}
\description{
Calculates average dry spell duration (below threshold)
}
