% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_avgDwellTime}
\alias{func_avgDwellTime}
\title{Calculates the average dwell time, i.e. average time for below median value spells}
\usage{
func_avgDwellTime(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates the average dwell time, i.e. average time for below median value spells
}
