% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_cor}
\alias{func_cor}
\title{Calculates the lag-1 autocorrelation}
\usage{
func_cor(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates the lag-1 autocorrelation
}
