% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_CSL}
\alias{func_CSL}
\title{Calculates the cold season length}
\usage{
func_CSL(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates the cold season length
}
