% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_cv}
\alias{func_cv}
\title{Calculates the coefficient of variation (mead/sd)}
\usage{
func_cv(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates the coefficient of variation (mead/sd)
}
