% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_dyWet}
\alias{func_dyWet}
\title{Calculates average rainfall on wet days (above threshold)}
\usage{
func_dyWet(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$threshold denoting the threshold}
}
\description{
Calculates average rainfall on wet days (above threshold)
}
