% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_F0}
\alias{func_F0}
\title{Calculates the number of frost days}
\usage{
func_F0(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates the number of frost days
}
