% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_GSL}
\alias{func_GSL}
\title{Calculates the growing season length}
\usage{
func_GSL(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates the growing season length
}
