% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_maxDSD}
\alias{func_maxDSD}
\title{Calculates maximum dry spell duration (below threshold)}
\usage{
func_maxDSD(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$threshold denoting the threshold}
}
\description{
Calculates maximum dry spell duration (below threshold)
}
