% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_maxWSD}
\alias{func_maxWSD}
\title{Calculates maximum wet spell duration (above threshold)}
\usage{
func_maxWSD(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$threshold denoting the threshold}
}
\description{
Calculates maximum wet spell duration (above threshold)
}
