% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_normP}
\alias{func_normP}
\title{Calculates normalised quantile (quantile divided by mean)}
\usage{
func_normP(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$quant denoting the probability of the quantile}
}
\description{
Calculates normalised quantile (quantile divided by mean)
}
