% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_P}
\alias{func_P}
\title{Calculates a quantile value}
\usage{
func_P(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$quant denoting the probability of the quantile}
}
\description{
Calculates a quantile value
}
