% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_R}
\alias{func_R}
\title{Calculates the number of days above a threshold (often used for temperature)}
\usage{
func_R(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$threshold denoting the threshold}
}
\description{
Calculates the number of days above a threshold (often used for temperature)
}
