% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_rng}
\alias{func_rng}
\title{Calculates the inter-quantile range}
\usage{
func_rng(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$lim denoting the probability limit width}
}
\description{
Calculates the inter-quantile range
}
