% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_seasRatio}
\alias{func_seasRatio}
\title{Calculates seasonality ratio}
\usage{
func_seasRatio(data, attArgs)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$indexSeas corresponding to season of interest}
}
\description{
Calculates seasonality ratio
}
