% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_tot}
\alias{func_tot}
\title{Calculates total of time series}
\usage{
func_tot(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates total of time series
}
