% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_WDcor}
\alias{func_WDcor}
\title{Calculates the lag-1 autocorrelation for wet days}
\usage{
func_WDcor(data)
}
\arguments{
\item{data}{is a vector, representing a time series}
}
\description{
Calculates the lag-1 autocorrelation for wet days
}
