% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_wettest6monPeakDay}
\alias{func_wettest6monPeakDay}
\title{Calculates the day of year corresponding to the wettest 6 months}
\usage{
func_wettest6monPeakDay(data, attArgs = NULL)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$doy denoting the day of year for each value in the time series}
}
\description{
Calculates the day of year corresponding to the wettest 6 months
}
