% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{func_wettest6monSeasRatio}
\alias{func_wettest6monSeasRatio}
\title{Calculates the ratio of wet season to dry season rainfall, based on wettest6monPeakDay}
\usage{
func_wettest6monSeasRatio(data, attArgs = NULL)
}
\arguments{
\item{data}{is a vector, representing a time series}

\item{attArgs}{is a list, with attArgs$doy denoting the day of year for each value in the time series}
}
\description{
Calculates the ratio of wet season to dry season rainfall, based on wettest6monPeakDay
}
