% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR4J_funcs.R
\name{GR4J_wrapper}
\alias{GR4J_wrapper}
\title{System model wrapper GR4J}
\usage{
GR4J_wrapper(data, systemArgs, metrics)
}
\arguments{
\item{data}{list; contains daily precipitation and PET to be used in GR4J, 
in a list with entries  \emph{times}, \emph{P} and optionally \emph{PET}.}

\item{systemArgs}{list; contains \code{Param} which is a vector of GR4J 
parameters (obtained from \code{calGR4J}),
\code{dates} which is a vector of dates, 
and \code{PET} which is a optional vector of potential transpiration 
(required if PET not included in \code{data})}

\item{metrics}{a vector of metric names (including 'meanQ' for 
mean daily flow, 'P99' and 'P25' for 99th and 25th percentile daily flows, 
and 'min3yr' for minimum 3-year total flow)}
}
\value{
A vector of metric values
}
\description{
\code{GR4J_wrapper} runs the GR4J model, using the airGR package, for a given 
set of climate inputs and parameter values and produces a set of runoff metrics
}
