% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{mvFunc_cor}
\alias{mvFunc_cor}
\title{Calculates the correlation between two time series}
\usage{
mvFunc_cor(data.1, data.2)
}
\arguments{
\item{data1}{and data2 are vectors, representing a time series}
}
\description{
Calculates the correlation between two time series
}
