% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{mvFunc_cvWetDay}
\alias{mvFunc_cvWetDay}
\title{Calculates the coefficient of variation (sdev/mean) value of a non-rainfall time series on wet-days}
\usage{
mvFunc_cvWetDay(data.1, data.2)
}
\arguments{
\item{data1}{represents a non-rainfall time series}

\item{data2}{represents rainfall time series}
}
\description{
Calculates the coefficient of variation (sdev/mean) value of a non-rainfall time series on wet-days
}
