% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{shuffle_sim}
\alias{shuffle_sim}
\title{Post-processing to apply changes in temporal structure of annual precipitation.}
\usage{
shuffle_sim(
  sim,
  clim,
  attPerturb = "P_day_all_tot_dwellTime",
  targetVals,
  targetType = "frac",
  seed = 1,
  annAR1coeffList = seq(-0.2, 0.9, 0.01),
  cSel = "mean"
)
}
\arguments{
\item{sim}{a list; simulated climate object from \code{generateScenarios}}

\item{clim}{a list; reference climate}

\item{attPerturb}{a string; name of attribute to be perturbed}

\item{targetVals}{a vector of numbers; targets for perturbed attributes}

\item{targetType}{a vector of strings; type of change in target attributes (either 'frac' or 'diff')}

\item{seed}{an integer; random number seed}

\item{annAR1coeffList}{a vector of numbers; the range of values for annual AR(1) parameters used in grid search}

\item{cSel}{an integer or string; for multi-site data, can either cvalculate perturbed attributes on a single site (integer) or the mean climate over all sites (\code{cSel='mean'}).}
}
\value{
A list with simulated perturbed climates. Same format as output from \code{generateScenarios}.
}
\description{
\code{shuffle_sim} changes the temporal structure of annual climate to match 
a target change in a parturbed attribute
}
