% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{viewAttributeDef}
\alias{viewAttributeDef}
\title{Prints the definition of an attribute}
\usage{
viewAttributeDef(attribute)
}
\arguments{
\item{attribute}{A string; the name of the attribute.}
}
\description{
\code{viewAttributeDef} prints the short definition of a valid attribute
}
\examples{
# To view the definition of any valid attribute
viewAttributeDef("P_day_all_tot_m")
}
\seealso{
\code{createExpSpace}
}
