% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{viewAttributes}
\alias{viewAttributes}
\title{Prints the names of valid attributes}
\usage{
viewAttributes()
}
\description{
\code{viewAttributes()} prints the names of valid attributes that may be used to create an exposure space.
}
\details{
The function does not take any input arguments. The valid attributes that may be specified as \code{attPerturb} or \code{attHold} to create an exposure space using the function \code{createExpSpace} are printed.
}
\examples{
# To view the valid attributes.
viewAttributes()

# To view the definition of any valid attribute
viewAttributeDef("P_ann_tot_m")
}
\seealso{
\code{createExpSpace}, \code{viewAttributeDef}
}
