% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{viewDefaultOptimArgs}
\alias{viewDefaultOptimArgs}
\title{Prints the default optimisation arguments}
\usage{
viewDefaultOptimArgs()
}
\description{
\code{viewDefautOptimArgs()} prints the default values of optimisation arguments (\code{optimisationArguments}) used by \code{generateScenarios}
}
\details{
The function does not take any input arguments. This a helper function that prints the default values of the optimisation arguments.
The user may specify alternate values of these arguments in fields named according to the corresponding argument name nested under 
\code{optimisationArguments} in a JSON file to use as the \code{controlFile} input to the \code{generateScenarios} function.
}
\examples{
# To view the default optimisation arguments
viewDefaultOptimArgs()
}
\seealso{
\code{writeControlFile}
}
