% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlFileFunctions.R
\name{viewModels}
\alias{viewModels}
\title{Prints the available stochastic model options}
\usage{
viewModels(variable = NULL, compatibleAtts = FALSE)
}
\arguments{
\item{variable}{String; the variable name. Type \code{viewModels()} without arguments to view the valid variable names.}

\item{compatibleAtts}{TRUE/FALSE (default is FALSE). Whether the attributes compatible with each model should be printed.}
}
\description{
\code{viewModels} prints the stochastic model options available for the different hydroclimatic variables in foreSIGHT.
                  These options may be used to create an controlFile for input to function \code{generateScenarios}.
}
\examples{
# To view the valid variable names use the function without arguments
viewModels()

# Examples to view the model options available for different variables
viewModels("P")
viewModels("Temp")
viewModels("Radn")
viewModels("PET")
}
\seealso{
\code{writeControlFile}, \code{generateScenarios}
}
