% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoTankModel.R
\name{viewTankMetrics}
\alias{viewTankMetrics}
\title{Prints the names of the performance metrics of the rain water tank system model}
\usage{
viewTankMetrics()
}
\description{
\code{viewTankMetrics} prints the names of the performance metrics available in the example rain water tank system model. T
}
\details{
This is a helper function that does not take any input arguments. The user may specify one or more of the metric names
as the \code{metric} argument of \code{tankWrapper} to select the performance metrics from the tank system model.
to select the performance metrics.
}
\examples{
viewTankMetrics()
}
\seealso{
\code{tankWrapper}
}
