% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{viewVariables}
\alias{viewVariables}
\title{Prints the names of and units of valid variables}
\usage{
viewVariables()
}
\description{
\code{viewVariables()} prints the names of valid variables and their units in the package. The user should input these variable
in the same units.
}
\details{
The function does not take any input arguments.
}
\examples{
# To view the valid variables
viewVariables()
}
\seealso{
\code{generateScenarios}
}
