% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_system_metrics.R
\name{boxplot_prob}
\alias{boxplot_prob}
\title{Draws a boxplot with the whiskers at specified probability limits}
\usage{
boxplot_prob(xin, whiskersProb = c(0.025, 0.975), at = NULL, ...)
}
\arguments{
\item{xin}{a vector, matrix or dataframe; data to be plotted}

\item{whiskersProb}{a vector of length 2; min and max probability limits}

\item{at}{a vector; specifying x coordinates for boxes}

\item{...}{; other arguments for \code{bxp}}
}
\value{
The function returns a boxplot figure.
}
\description{
\code{boxplot_prob} Draws a boxplot with the whiskers at the probability limits, provided by \code{whiskersProb}.
}
