% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInputChecker.R
\name{convert_climYMD_POSIXct}
\alias{convert_climYMD_POSIXct}
\title{Converts "old" reference climate data format (<V1.2) to "new" format with POSIXct dates.}
\usage{
convert_climYMD_POSIXct(clim)
}
\arguments{
\item{clim}{data.frame or list; contains reference daily climate data from foreSIGHT V1.2 or earlier. \cr}
}
\value{
The function returns a list containing \code{times} in POSIXct format (corresponding to year, month and day from original data \code{clim}), and climate variables.
}
\description{
\code{convert_climYMD_POSIXct} produces reference climate data list for use in V2.0 and newer.
}
