% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egClimData}
\alias{egClimData}
\title{Climate attributes from projections.}
\format{
A data frame with 6 rows and 6 variables:
\describe{
  \item{P_ann_tot_m}{change in mean annual total P, fraction}
  \item{P_ann_seasRatio_m}{change in seasonal ratio of P, fraction}
  \item{P_ann_nWet_m}{change in the number of wet days, fraction}
  \item{Temp_ann_avg_m}{change in average annual Temp, additive}
  \item{Name}{name of the climate model}
  \item{Avg. Deficit}{performance metric values}
}
}
\usage{
egClimData
}
\description{
A example dataset containing the climate attribute values 
in fraction/additive change
}
\keyword{datasets}
