% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egScalPerformance}
\alias{egScalPerformance}
\title{Performance metrics of the tank model using simple scaled scenarios.}
\format{
A list with 2 elements
\describe{
  \item{Avg. Deficit}{average daily deficit of water, litres}
  \item{Reliability}{reliability of the tank, fraction}
}
}
\usage{
egScalPerformance
}
\description{
Performance metrics of the tank model using simple scaled scenarios.
}
\keyword{datasets}
