% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{data}
\name{egScalPerformance}
\alias{egScalPerformance}
\title{Performance metrics of the tank model using simple scaled scenarios.}
\format{
A list with 2 elements
\describe{
  \item{volumetric reliability (fraction)}{Volumetric reliaiblity of tank}
  \item{reliability (fraction)}{Reliability of tank}
}
}
\usage{
egScalPerformance
}
\description{
Performance metrics of the tank model using simple scaled scenarios.
}
\keyword{datasets}
