% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_system_metrics.R
\name{evaluate_system_metrics}
\alias{evaluate_system_metrics}
\title{Calculates system metrics for and observed and baseline stochastic climates}
\usage{
evaluate_system_metrics(
  sim,
  clim,
  systemModel,
  systemArgs,
  metrics,
  varNames = NULL
)
}
\arguments{
\item{sim}{list; a simulation containing the scenarios generated using the function \code{generateScenarios}.}

\item{clim}{a list; reference climate \cr}

\item{systemModel}{a function; The function runs the system model using climate data in a list as input.
The function is expected to be created by the user for specific system models.}

\item{systemArgs}{a list; containing the input arguments to \code{systemModel}.}

\item{metrics}{a string vector; the names of the performance metrics the \code{systemModel} function returns.}

\item{varNames}{a string vector; containing the names of the climate variables that are extracted from sim and used in system model. If \code{NULL}, then \code{varNames} determined from attribute names in \code{sim$expSpace}.}
}
\value{
a list containing \code{systemPerf_base} and \code{systemPerf_obsClim}, with performance metrics for baseline stochastic climate and observed climate, respectively.
}
\description{
\code{evaluate_system_metrics} runs observed climate and baseline
(unperturbed) stochastic climates through a system model and calculates system metrics for each.
This is used to perform evaluation of stochastic climates using the 'virtual observation' approach.
}
\examples{
# load dates, precip, PET and streamflow data for Scott Creek
data('data_A5030502')
clim_ref = list(times = data_A5030502$times,P = data_A5030502$P)  
data("egScottCreekSimStoch")
# observed flow
Qobs = data_A5030502$Qobs
# observed PET 
PET = data_A5030502$PET 
dates = as.Date(clim_ref$times)
# calibrate GR4J parameters
Param = calGR4J(dates = dates,P=clim_ref$P,PET=PET,Qobs=Qobs)
# setup systemArgs and metrics
systemArgs = list(dates=dates,Param=Param,PET=PET)
metrics = c('meanQ','P99','P25','min3yr')
eval = evaluate_system_metrics(sim=sim.stoch,clim=clim_ref,
                               systemModel=GR4J_wrapper,systemArgs=systemArgs,
                               metrics=metrics)
par(mfrow=c(2,2),mar=c(3,4,1,1))
for (m in 1:length(metrics)){
  metric = metrics[m]
  yAll = c(eval$systemPerf_base[[metric]],eval$systemPerf_obsClim[metric])
  ylim = c(min(yAll),max(yAll))
  boxplot_prob(as.vector(eval$systemPerf_base[[metric]]),ylim=ylim)
  points(eval$systemPerf_obsClim[metric],col='red',pch=19)
  title(metric)
}
}
