% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreSIGHT.R
\docType{package}
\name{foreSIGHT}
\alias{foreSIGHT}
\title{foreSIGHT: A package for Systems Insights from Generation of Hydroclimatic Timeseries}
\description{
A tool to create hydroclimate scenarios, stress test systems and visualize
system performance in scenario-neutral climate change impact assessments.
}
