% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foresight.R
\docType{package}
\name{foresight}
\alias{foresight}
\title{foresight}
\description{
\itemize{
\item Online App:
\item Github: https://github.com/JGCRI/foresight
\item Webpage: https://jgcri.github.io/foresight/}
}
