% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{getSimSummary}
\alias{getSimSummary}
\title{Produces a summary object containing the metadata of a full simulation}
\usage{
getSimSummary(sim)
}
\arguments{
\item{sim}{list; a simulation containing the scenarios generated using the function \code{generateScenarios}.}
}
\description{
\code{getSimSummary} uses a full simulation generated using the function \code{generateScenarios} as input and outputs a summary object containing the
metadata of the full simulation. The output summary object may be used as an input to the plotting functions in this package. The output summary object
will be much smaller in size than the full simulation for ease of storage and use with the plotting functions.
}
\seealso{
\code{generateScenarios}, \code{plotPerformanceSpace}, \code{plotPerformanceOAT}
}
