% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{mvFunc_avgDryDay}
\alias{mvFunc_avgDryDay}
\title{Calculates the average value of a non-rainfall time series on dry-days}
\usage{
mvFunc_avgDryDay(data.1, data.2)
}
\arguments{
\item{data.1}{represents a non-rainfall time series}

\item{data.2}{represents rainfall time series}
}
\description{
Calculates the average value of a non-rainfall time series on dry-days
}
