% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{mvFunc_avgWetDay}
\alias{mvFunc_avgWetDay}
\title{Calculates the average value of a non-rainfall time series on wet-days}
\usage{
mvFunc_avgWetDay(data.1, data.2)
}
\arguments{
\item{data.1}{represents a non-rainfall time series}

\item{data.2}{represents rainfall time series}
}
\description{
Calculates the average value of a non-rainfall time series on wet-days
}
