% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeManager.R
\name{mvFunc_cor}
\alias{mvFunc_cor}
\title{Calculates the correlation between two time series}
\usage{
mvFunc_cor(data.1, data.2)
}
\arguments{
\item{data.1}{a vector for the first climate variable}

\item{data.2}{a vector for the second climate variable}
}
\description{
Calculates the correlation between two time series
}
