% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpSpace.R
\name{plotExpSpace}
\alias{plotExpSpace}
\title{Plots the location of points in a two-dimensional exposure space}
\usage{
plotExpSpace(
  expSpace,
  y = expSpace[["attPerturb"]][1],
  x = expSpace[["attPerturb"]][2]
)
}
\arguments{
\item{expSpace}{list; an exposure space created using the function \code{createExpSpace}}

\item{y}{a string; tag of a perturbed attribute to plot on the y-axis. Defaults to \code{expSpace[["attPerturb"]][1]}.}

\item{x}{a string; tag of a perturbed attribute to plot on the x-axis. Defaults to \code{expSpace[["attPerturb"]][2]}.}
}
\description{
The function uses an exposure space created using the function \code{createExpSpace} as input and creates a plot of the
two dimensional (2D) exposure space. \code{plotExpSpace} plots only 2D spaces consisting of samples of 2 attributes.
}
\details{
The number of dimensions of an exposure space is equal to the number of perturbed attributes in that space. If the
exposure space has more than 2 dimensions (perturbed attributes), this function can be used to plot 2D slices of the space.
Note that the default arguments of this function is defined to plot a slice showing the first two dimensions of the space,
arguments \code{x} and \code{y} may be specified to plot alternate dimensions.
}
\examples{
# create an exposure space that has more than 2 dimensions
attPerturb <- c("P_day_all_tot_m", "P_day_all_nWet_m", "P_day_Feb_tot_m")
attHold <- c(
  "P_day_SON_dyWet_m", "P_day_JJA_avgWSD_m", "P_day_MAM_tot_m", "P_day_DJF_avgDSD_m",
  "Temp_day_all_rng_m", "Temp_day_all_avg_m"
)
attPerturbType <- "regGrid"
attPerturbSamp <- c(5, 5, 5)
attPerturbMin <- c(0.8, 0.9, 0.85)
attPerturbMax <- c(1, 1.1, 1.05)
expSpace <- createExpSpace(
  attPerturb = attPerturb,
  attPerturbSamp = attPerturbSamp,
  attPerturbMin = attPerturbMin,
  attPerturbMax = attPerturbMax,
  attPerturbType = attPerturbType,
  attHold = attHold,
  attTargetsFile = NULL
)
# plot the first two dimensions
plotExpSpace(expSpace)
# plot another slice
plotExpSpace(expSpace, y = "P_day_all_tot_m", x = "P_day_Feb_tot_m")
}
\seealso{
\code{createExpSpace}
}
