% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_attribute_change.R
\name{plotPerformanceAttributesOAT}
\alias{plotPerformanceAttributesOAT}
\title{Plots changes in attributes for a specified perturbed attribute}
\usage{
plotPerformanceAttributesOAT(
  clim,
  sim,
  attPerturb,
  attEval,
  vSel = NULL,
  cSel = NULL,
  ylim = NULL,
  cex.main = 0.8,
  cex.xaxis = 0.5,
  cex.yaxis = 0.5
)
}
\arguments{
\item{clim}{list; reference climate data. \cr}

\item{sim}{list; perturbed climate from \code{generateScenarios()} \cr}

\item{attPerturb}{string; name of perturbed attribute \cr}

\item{attEval}{string; name of attribute that will be evaluated \cr}

\item{vSel}{string; variable name for selection site/aggregation \cr}

\item{cSel}{integer or string 'mean'; how to summarize multisite data - select site number or 'mean' for average \cr}

\item{ylim}{numeric vector of length 2; min and max y limits for plotting \cr}

\item{cex.main}{number; size for title \cr}

\item{cex.xaxis}{number; size for x-axis \cr}

\item{cex.yaxis}{number; size for y-axis \cr}
}
\value{
The function returns a single plot showing changes in attribute \code{attEval} for changes in perturbed attribute \code{attPerturb} \cr
}
\description{
\code{plotPerformanceAttributesOAT} plots OAT changes in attributes based on simulated climate object, for a given perturbed attribute.
}
\examples{
# XXXXXXXXXX
}
